LayoutTests/imported/w3c/web-platform-tests/mathml/relations/css-styling/lengths-2-expected.txt


FAIL Units assert_equals: q expected 192 but got 0
FAIL Trimming of space assert_equals: q expected 192 but got 0
PASS Non-negative numbers
PASS Non-positive numbers
PASS Legacy namedspaces
FAIL Legacy numbers assert_approx_equals: n3 expected 0 +/- 0.5 but got 70
